[WIP] Refactor to use encode_calldata everywhere we need to call functions
