+++ cmd/ethrex_l2/Cargo.toml
@@ -27,6 +27,7 @@ secp256k1.workspace = true
 keccak-hash = "0.10.0"
 
 ethrex-l2.workspace = true
+ethrex-sdk.workspace = true
 ethrex-core.workspace = true
 ethrex-blockchain.workspace = true
 ethrex-prover.workspace = true
