+++ cmd/ethrex_l2/src/commands/info.rs
@@ -1,7 +1,7 @@
 use crate::config::EthrexL2Config;
 use clap::Subcommand;
 use colored::{self, Colorize};
-use ethrex_l2::utils::eth_client::EthClient;
+use ethrex_l2_sdk::eth_client::EthClient;
 use keccak_hash::H256;
 use std::str::FromStr;
