+++ cmd/ethrex_l2/src/commands/test.rs
@@ -3,7 +3,7 @@ use bytes::Bytes;
 use clap::Subcommand;
 use ethereum_types::{Address, H160, H256, U256};
 use ethrex_blockchain::constants::TX_GAS_COST;
-use ethrex_l2::utils::eth_client::{eth_sender::Overrides, EthClient};
+use ethrex_l2_sdk::eth_client::{eth_sender::Overrides, EthClient};
 use keccak_hash::keccak;
 use secp256k1::SecretKey;
 use std::{
