+++ crates/l2/proposer/errors.rs
@@ -1,10 +1,11 @@
 use std::sync::mpsc::SendError;
 
-use crate::utils::merkle_tree::MerkleError;
-use crate::utils::{config::errors::ConfigError, eth_client::errors::EthClientError};
+use crate::utils::config::errors::ConfigError;
 use ethereum_types::FromStrRadixErr;
 use ethrex_core::types::{BlobsBundleError, FakeExponentialError};
 use ethrex_dev::utils::engine_client::errors::EngineClientError;
+use ethrex_l2_sdk::eth_client::errors::EthClientError;
+use ethrex_l2_sdk::merkle_tree::MerkleError;
 use ethrex_storage::error::StoreError;
 use ethrex_vm::EvmError;
 use tokio::task::JoinError;
