+++ crates/l2/sdk/Cargo.toml
@@ -4,7 +4,6 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
-ethrex-l2.workspace = true
 ethrex-core.workspace = true
 ethrex-rpc.workspace = true
 
@@ -15,6 +14,14 @@ keccak-hash = "0.11.0"
 secp256k1.workspace = true
 itertools = "0.13.0"
 thiserror.workspace = true
+serde_json.workspace = true
+serde.workspace = true
+tracing.workspace = true
+ethrex-rlp.workspace = true
+bytes.workspace = true
+reqwest.workspace = true
+
+eyre = "0.6"
 
 [lib]
 name = "ethrex_l2_sdk"
