+++ crates/l2/utils/config/committer.rs
@@ -1,5 +1,5 @@
-use crate::utils::secret_key_deserializer;
 use ethereum_types::Address;
+use ethrex_l2_sdk::secret_key_deserializer;
 use secp256k1::SecretKey;
 use serde::Deserialize;
